requirements.txt
customtkinter==5.2.1
pillow==10.2.0
pygame==2.5.2
